:recycle: 优化build软连接
